backend/pyproject.toml
[tool.poetry]
name = "housing-case-navigator"
version = "0.1.0"
description = ""
authors = ["tom11my <tommy11jo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
openai = "^1.54.3"
anthropic = "^0.39.0"
pytesseract = "^0.3.13"
pymupdf = "^1.24.13"
fastapi = "^0.115.4"
python-multipart = "^0.0.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
